+++ frontend/pages/hosts/ManageHostsPage/HostTableConfig.tsx
@@ -73,14 +73,6 @@ interface IDeviceUserCellProps {
     original: IHost;
   };
 }
-interface IDiskSpaceCellProps {
-  cell: {
-    value: any;
-  };
-  row: {
-    original: IHost;
-  };
-}
 
 const condenseDeviceUsers = (users: IDeviceUser[]): string[] => {
   if (!users?.length) {
@@ -195,7 +187,7 @@ const allHostTableHeaders: IDataColumn[] = [
     ),
   },
   {
-    title: "Disk space",
+    title: "Disk space available",
     Header: (cellProps: IHeaderProps) => (
       <HeaderCell
         value={cellProps.column.title}
@@ -205,28 +197,80 @@ const allHostTableHeaders: IDataColumn[] = [
     accessor: "gigs_disk_space_available",
     Cell: (cellProps: INumberCellProps): JSX.Element => {
       const { id, percent_disk_space_available } = cellProps.row.original;
+
+      const diskSpaceTooltip = () => {
+        const diskSpaceAvailable = cellProps.cell.value;
+        switch (true) {
+          case diskSpaceAvailable < 16:
+            return (
+              <span className={`tooltip-text`}>
+                Not enough disk space <br />
+                available to install most <br />
+                small operating systems <br />
+                updates.
+              </span>
+            );
+          case diskSpaceAvailable < 32:
+            return (
+              <span className={`tooltip-text`}>
+                Not enough disk space <br />
+                available to install most <br />
+                large operating systems <br />
+                updates.
+              </span>
+            );
+          default:
+            return (
+              <span className={`tooltip-text`}>
+                Enough disk space available <br />
+                to install most operating <br />
+                systems updates.
+              </span>
+            );
+        }
+      };
+
+      const diskSpaceIndicator = () => {
+        const diskSpaceAvailable = cellProps.cell.value;
+        switch (true) {
+          case diskSpaceAvailable < 16:
+            return "red";
+          case diskSpaceAvailable < 32:
+            return "yellow";
+          default:
+            return "green";
+        }
+      };
+
       if (cellProps.cell.value === 0) {
         return <>No data available</>;
       }
       return (
         <>
-          <span
-            className={`graph-cell`}
+          <div
+            className="gigs_disk_space_available__cell__disk-space-wrapper"
             data-tip
-            data-for={`disk-space__${cellProps.row.original.id}`}
+            data-for={`disk-space__${id}`}
           >
-            Graph goes here
-          </span>
+            <div className="gigs_disk_space_available__cell__disk-space">
+              <div
+                className={`gigs_disk_space_available__cell__disk-space-${diskSpaceIndicator()}`}
+                style={{
+                  width: `${100 - percent_disk_space_available}%`,
+                }}
+              />
+            </div>
+          </div>
           <ReactTooltip
-            place="top"
+            place="bottom"
             type="dark"
             effect="solid"
             backgroundColor="#3e4771"
-            id={`device_mapping__${cellProps.row.original.id}`}
+            id={`disk-space__${id}`}
             data-html
           >
             <span className={`tooltip__tooltip-text`}>
-              Tooltip text goes here
+              {diskSpaceTooltip()}
             </span>
           </ReactTooltip>{" "}
           <span>{cellProps.cell.value} GB</span>
